Add __FILE__ which gets the path name of the compilation unit.
Note however that to avoid overhead in the compilation code, __FILE__ change
its value with every call to 'import'. So the only safe usage is to copy the
value in a unique variable and use it afterwards with the expected results.

It has to be a unique name because the language doesn't have the capability for
static or local variables with scope to the compilation unit, for that it needs
namespaces.

Add -g to the normal build CFLAGS.