This is to merge a major upstream development, which extends possibilities, like to be able to run multiply virtual machine instances/states, without interfering with each other.
Such a state now, is passing arround as the first function argument.

That was an impressive work by Jason, so many thanks.

In this commit though, the len () method to the core datatypes hasn't been implement,
as there is an uncertainity if it is rather better to stay as is as a native function,
though feels more natural when is attached as a method and looks consistent.

The thing is that for StringType objects, len() is clearly ambiguous, and so it has
to be implemented a bytelen() for byte semantics, and a charlen() to mean character
semantics.