Add isDefined() native function, that returns either true (when given object has been defined) or false if not.
Note that, this performs a lookup only to the global table for object references, as
this is the common use of a such a function. For local objects this feels unecessary.

Anyway, previously this function, was a method of the String Class, and while indeed
is a very compact way to express, is not actually a StringType function, so it makes
sence as a native function.

Note and with regards to the usage case and besides the obvious when you really want
to know if (lets say a class) is defined:

Maybe there are some cases, when you want uniq named object, so it cannot be shadowed
by another in the chain. Currently there is an open ticket to the upstream tracker, to
introduce 'const'ant type objects, which after initialization the assigned value can
not be change.
Here we are talking for a bit different thing, which simply the desired object, can
not be used outside of its compilation scope.